feat: defining clado version 0.0.1 package
