--- package.json
@@ -1,50 +1,0 @@
-{
-  "private": false,
-  "name": "clado",
-  "version": "1.0.0",
-  "description": "A React state machine based on Context API.",
-  "main": "dist/index.js",
-  "scripts": {
-    "lint": "prettier --check .",
-    "lint:fix": "prettier --write .",
-    "build": "./scripts/build.sh",
-    "test": "./scripts/test.sh",
-    "test:ci": "./scripts/test-ci.sh",
-    "publish": "./scripts/publish.sh"
-  },
-  "keywords": [
-    "state",
-    "machine",
-    "context",
-    "react",
-    "component",
-    "clade",
-    "komutilo"
-  ],
-  "author": "Bruno Araujo <bruno@lavaleakgames.com>",
-  "license": "MIT",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/komutilo/clado.git"
-  },
-  "devDependencies": {
-    "@babel/core": "^7.14.5",
-    "@babel/preset-env": "^7.14.5",
-    "@babel/preset-react": "^7.14.5",
-    "@babel/preset-typescript": "^7.14.5",
-    "@testing-library/jest-dom": "^5.14.1",
-    "@testing-library/react": "^11.2.7",
-    "@testing-library/react-hooks": "^7.0.0",
-    "@testing-library/user-event": "^13.1.9",
-    "@types/jest": "^26.0.23",
-    "@types/node": "^15.12.2",
-    "@types/react": "^17.0.11",
-    "@types/react-dom": "^17.0.7",
-    "jest": "^27.0.4",
-    "jest-junit": "^12.2.0",
-    "prettier": "^2.3.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "typescript": "^4.3.2"
-  }
-}
