+++ packages/clado/.jest/setup.js
@@ -1,0 +1,2 @@
+import '@testing-library/jest-dom';
+import '@testing-library/jest-dom/jest-globals';
