+++ packages/clado/.prettierignore
@@ -1,4 +1,4 @@
 **/node_modules
 /dist
 /coverage
-yarn.lock
+pnpm-lock.yaml
