+++ packages/clado/package.json
@@ -1,0 +1,62 @@
+{
+  "name": "clado",
+  "version": "0.0.1",
+  "description": "A React state machine based on Context API.",
+  "main": "dist/index.js",
+  "keywords": [
+    "state",
+    "machine",
+    "context",
+    "react",
+    "component",
+    "clade",
+    "komutilo"
+  ],
+  "author": "Bruno Araujo <bruno@lavaleakgames.com>",
+  "license": "MIT",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/komutilo/clado.git"
+  },
+  "engines": {
+    "node": ">=18.12.0",
+    "npm": "8.19.4"
+  },
+  "volta": {
+    "node": "18.12.0",
+    "npm": "8.19.4",
+    "pnpm": "8.15.1",
+    "yarn": "4.1.0"
+  },
+  "scripts": {
+    "lint": "prettier --check .",
+    "lint:fix": "prettier --write .",
+    "build": "tsc",
+    "test": "jest --forceExit --detectOpenHandles --env=jsdom --watch",
+    "test:ci": "jest --forceExit --detectOpenHandles --env=jsdom --reporters=default --maxWorkers=1 --reporters=jest-junit --coverage --watchAll=false"
+  },
+  "devDependencies": {
+    "@babel/core": "^7.23.9",
+    "@babel/preset-env": "^7.23.9",
+    "@babel/preset-react": "^7.23.3",
+    "@babel/preset-typescript": "^7.23.3",
+    "@testing-library/jest-dom": "^6.4.2",
+    "@testing-library/react": "^14.2.1",
+    "@testing-library/user-event": "^14.5.2",
+    "@types/jest": "^29.5.12",
+    "@types/node": "^20.11.19",
+    "@types/react": "^18.2.56",
+    "@types/react-dom": "^18.2.19",
+    "jest": "^29.7.0",
+    "jest-environment-jsdom": "^29.7.0",
+    "jest-junit": "^16.0.0",
+    "prettier": "^3.2.5",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "rimraf": "^5.0.5",
+    "typescript": "^5.3.3"
+  },
+  "peerDependencies": {
+    "react": "^16.8.0"
+  }
+}
