--- scripts/build.sh
@@ -1,4 +1,0 @@
-#!/bin/bash
-
-test -d dist && rm -rf dist
-tsc
