--- scripts/test.sh
@@ -1,7 +1,0 @@
-#!/bin/bash
-
-jest \
-    --forceExit \
-    --detectOpenHandles \
-    --env=jsdom \
-    --watch
